Improved constant expansion for ARM
In order to avoid some problems with code size we don't want to emit the
`movt/movw` instruction when code size plays a role.

Furthermore we can remove `literals_in_code` since it always `true` and
if it would be `false` it would also use `movt/movw` and thus only work
if the target supports Thumb2.