Add `Tbool` case to `retttype`
So that builtins and external functions with return type `_Bool` can
be described more precisely.

Values returned by a function of return type `_Bool` are still normalized
as 8-bit integers, because that's what the x86 and AArch64 ABI say
(the return value is an 8-bit integer that is equal to 0 or to 1).