Clean up the definition of `classify_cast`
Make casts to `_Bool` a case of casts to integer types, rather than a
special case at the beginning.

Clarify the role of `Archi.ptr64` in casts from pointer types to int and long.

The new definition of `classify_cast` was proved equivalent to the old
definition, so there is no change in the semantics of CompCert C and Clight.