Value analysis: take types of function parameters into account
We know that the values of function parameters match the types declared
in the function signature.  We can use this information to give more
precise initial abstract values for parameters of small integer types.

As an exception: if a parameter needs normalization on function entry
(as per the ABI), ignore its type, otherwise the normalization code
could be optimized away.