More conservative analysis of pointer equality and difference
A pointer "one past" a global "x" and a pointer to the beginning of a
global "y" can be equal.  This is undefined behavior in CompCert C,
but defined behavior in ISO C.

Thus, the value analysis should not infer that a pointer based at "x"
and a pointer based at "y" cannot be equal.

The only inference we keep is that a pointer based in the current
stack block and a pointer based in a global variable cannot be equal.