Check that preprocessed source do not contain backslash-newline
This is normally ensured by the preprocessor, as translation phases 1 and 2
(ISO C99, 5.1.1.2).  However, these sequences could occur in
hand-written .i files and confuse the CompCert lexer.

Backslash-newline can also be written using trigraphs: `??/`-newline.
An occurrence of this sequence in a preprocessed file is highly suspicious,
so we reject it as well.