Replace `C2C.atom_is_extern` by more robust criterion `C2C.atom_is_external`
This predicate is used in macOS and Cygwin ports to determine when a global symbol may be defined in a shared library, and therefore must be referenced through the GOT.

The previous criterion was just "is it declared `extern`"?  However, this misses some cases, e.g. "common" declaration.

The new criterion is "it is not declared `static` and not defined in the current compilation unit", which should be a lot more robust.

Fixes: #537