Distinguish more clearly the warnings that apply to extracted OCaml files
In GNU make, a target-dependent variable assignment such as
       pattern: VAR += extra
triggers not only when the file being remade matches the pattern, but
also when this file is a prerequisite of another file that matches the
pattern.

This caused hand-written OCaml files to be compiled twice with
different warning options:
- the options for extracted files when the compilation is triggered by
  a dependency from an extracted file;
- the other set of options otherwise.

Different .cmi can be generated when compiling with/without some
warnings (e.g. warning 56), so this can trigger link-time
incompatibilities.

This commit distinguishes more clearly the warnings that apply to
extracted OCaml files (in extraction/) from those that apply to
hand-written files (elsewhere).