Revised handling of unknown warnings (#554)
`-W<some unknown warning>` is now (by default) a warning (not a fatal error)
and `-Wno-<some unknown warning>` is ignored.
This is what GCC and Clang do.

This commit also introduces the `unknown-warning-option` warning, so
that warnings about unknown warnings can be ignored
(`-Wno-unknown-warning-option`) or turned into an error
(`-Werror=unknown-warning-option`).  This is what Clang does.

Closes: #552