Small fixes for unknown-warning-option.
Using `Ignore` causes the compiler to ignore the next command line
option, we shoud use `Unit` with an empty function.

Also we need to check if there was any error during parsing the
commandline for checking if an we have an unknown warning and the
warning is set to an error.
Bug 39421