Pass type of controlling switch expression into ctx.
The standard requires that the constant expression of a case needs to be
converted to the integer promoted type of the controlling expression
(§6.8.4.2p5).

To enforce this we pass the ineger promoted type of the controlling
expression in ctx  and add a cast of the case expression to this type.

If the controlling expression is not an integer type we already have an
error message and skip adding the cast.
Bug 40122